test(EXUI-4513): move header scenarios to helper
